:bug: bug: 이메일 알림 발송 로직 수정 중 발생한 오류 해결 완료
- 키워드가 많이 겹치는 순으로 내림차순 정렬이 제대로 반영 안 되는 문제 해결: limit(20)을 두 번째 쿼리에 넣고, 사용자 정보를 먼저 조회하도록 순서를 바꿈
